# web/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Robot Operation Center (ROC). Internal automation platform to distribute and execute robots across different subnets.",
    "version": "1.0.0",
    "title": "ROC (Robot Operation Center)",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/agent": {
      "get": {
        "operationId": "getAgents",
        "summary": "List all agents by project id",
        "tags": [
          "agent"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "project object to be returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createAgent",
        "summary": "Create new agent",
        "tags": [
          "agent"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Agent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation. Returns an agent object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      }
    },
    "/agent/{agentId}": {
      "get": {
        "operationId": "getAgentById",
        "summary": "Get agent by id",
        "tags": [
          "agent"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "description": "Id of agent",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation. Returns an agent object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateAgent",
        "summary": "Update Agent State",
        "tags": [
          "agent"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "description": "Id of agent",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation. Returns an agent object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Agent"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteAgent",
        "summary": "Delete Agent",
        "tags": [
          "agent"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "agentId",
            "description": "Id of agent",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation."
          }
        }
      }
    },
    "/environment": {
      "get": {
        "operationId": "getEnvironments",
        "summary": "List all environments by project id",
        "tags": [
          "environment"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "description": "Id of project",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "project object to be returned",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Environment"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createEnvironment",
        "summary": "Create new environment object",
        "tags": [
          "environment"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "projectId",
            "description": "Id of project",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Environment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation. Returns an environment object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Environment"
                }
              }
            }
          }
        }
      }
    },
    "/environment/{environmentId}": {
      "get": {
        "operationId": "getEnvironmentById",
        "summary": "Get environment by id",
        "tags": [
          "environment"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "environmentId",
            "description": "Id of environment",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation. Returns an environment object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Environment"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateEnvironment",
        "summary": "Update an environment object",
        "tags": [
          "environment"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "environmentId",
            "description": "Id of Environment",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation. Returns an environment object.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Environment"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteEnvironment",
        "summary": "Delete Environment",
        "tags": [
          "environment"
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "environmentId",
            "description": "Id of environment",
            "required": true,
            "schema": {
              "type": "string",
              "items": {
                "type": "integer"
              }
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful operation."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AccessTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Access-Token",
        "description": "Provide access token to perform operations."
      }
    },
...
    "schemas": {
...
      "Agent": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "platform": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "dockerVersion": {
            "type": "string"
          },
          "lastSeen": {
            "type": "string"
          }
        }
      },
      "Environment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}